Refactor down Gherkin::Messages::* to Gherkin::*
