+++ gherkin/ruby/spec/gherkin/gherkin_spec.rb
@@ -1,0 +1,15 @@
+require 'rspec'
+require 'gherkin/gherkin'
+
+module Gherkin
+  describe Gherkin do
+    it "works" do
+      cucumber_messages = Gherkin.new(
+        ["testdata/good/minimal.feature"],
+        true, true, true
+      )
+      messages = cucumber_messages.messages.to_a
+      expect(messages.length).to eq(3)
+    end
+  end
+end
